Hotfix for 24 hour return time.
Suntimes on certain dates at certain locations returned an hour of 24 in the DateTime value. C# will throw an argument exception during this event. anything beyond 24:00 will set the rise/set times for that date null. The update will result 24 int values to 0 to avoid this exception. The day will remain the same however, as return values must always be for the same Z day.